+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Added setting for putting the media query style block to <body> vs <head>
+- Debug mode. Throw debug exceptions only if debug is active.
+  ([#392](https://github.com/jjriv/emogrifier/pull/392))
 
 ### Changed
 
@@ -18,12 +20,34 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ### Removed
+- Drop support for HHVM
+  ([#386](https://github.com/jjriv/emogrifier/pull/386))
 
 
 ### Fixed
 
 
 ### Security
+
+
+
+## 1.2.0 (2017-03-02)
+
+### Added
+- Handling invalid xPath expression warnings
+  ([#361](https://github.com/jjriv/emogrifier/pull/361))
+
+
+### Deprecated
+- Support for PHP 5.5 will be removed in Emogrifier 3.0.
+- Support for PHP 5.4 will be removed in Emogrifier 2.0.
+
+
+### Fixed
+- Allow colon (:) and semi-colon (;) when using the *= selector
+  ([#371](https://github.com/jjriv/emogrifier/pull/371))
+- Ignore "auto" width and height
+  ([#365](https://github.com/jjriv/emogrifier/pull/365))